update example to latest react native version
