+++ Example/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'Example'
-
-include ':app'
 include ':react-native-bluetooth-state'
 project(':react-native-bluetooth-state').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-bluetooth-state/android')
+
+include ':app'
