+++ Example/package.json
@@ -1,17 +1,24 @@
 {
-  "name": "Example",
+  "name": "example",
   "version": "0.0.1",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
-    "postinstall": "rnpm link"
+    "test": "jest",
+    "postinstall": "react-native link react-native-bluetooth-state"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.25.1",
-    "react-native-bluetooth-state": "file:./.."
+    "react": "16.6.3",
+    "react-native": "0.57.8",
+    "react-native-bluetooth-state": "file:../"
   },
   "devDependencies": {
-    "rnpm": "^1.7.0"
+    "babel-jest": "23.6.0",
+    "jest": "23.6.0",
+    "metro-react-native-babel-preset": "0.48.5",
+    "react-test-renderer": "16.6.3"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
